rfnoc: Add context handling cores
The context_parser and context_builder cores handle the job of
interpreting and generating the context infromation for the raw data
interface of the NoC shell so that the user does not have to. The
context information (packet flags and timestamp) are repalced by
sideband data that moves with the data AXI stream.