rfnoc: Update error timestamp handling
This changes the error reporting so that the timestamp for the error
goes into the timestamp field of the control packet instead of being
passed in the payload of the packet. This allows us to use a single
packet for error reporting more easily, since the NoC shell doesn't
support master block writes.